git-svn-id: http://springhead.info/spr2/Springhead2/trunk@2176 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Physics/PHSolid.h
@@ -192,7 +192,7 @@ public:
 	Vec3d		GetOldAngularVelocity() const {return oldAngVel;}
 
 	///	”CˆÓ‚ÌˆÊ’u‚Å‚Ì‘¬“x‚ÌŽæ“¾
-	Vec3d		GetPointVelocity(Vec3f posW) const {
+	Vec3d		GetPointVelocity(Vec3d posW) const {
 		return velocity + (angVelocity^(posW - pose*center));
 	}
